(~) move InsertSliderValue into Component.Parametric partial — merge with main pushed Component.cs to 501 lines (500-line gate) (#825)
